requirements.txt
boto3==1.16.47
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
mypy-extensions==0.4.3
mypy==0.782
psycopg2-binary==2.8.6
pytest-cov==2.10.1
pytest==6.0.1
pytz==2020.5
selenium==3.141.0
slackclient==2.9.3
typing==3.7.4.3